# dashboard/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Numeraire SwissKnife REST API",
    "description": "This API is available to anyone with a Numeraire account. The `Me` (`/me`) endpoints expose the authenticated account and its wallet-scoped operations.",
    "contact": {
      "name": "Dario Anongba Varela",
      "email": "dario.varela@numeraire.tech"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/bitcoin/addresses": {
      "get": {
        "tags": [
          "Bitcoin Addresses"
        ],
        "summary": "List Bitcoin addresses",
        "description": "Returns all the Bitcoin addresses given a filter",
        "operationId": "list_btc_addresses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Total amount of results to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset where to start returning results",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "ids",
            "in": "query",
            "description": "List of IDs",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "wallet_id",
            "in": "query",
            "description": "Wallet ID. Automatically populated with your ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "Address",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "address_type",
            "in": "query",
            "description": "Status",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/BtcAddressType"
                }
              ]
            }
          },
          {
            "name": "used",
            "in": "query",
            "description": "Whether the address has been used",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "order_direction",
            "in": "query",
            "description": "Direction of the ordering of results",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BtcAddress"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bitcoin Addresses"
        ],
        "summary": "Generate a new Bitcoin address",
        "description": "Returns the generated Bitcoin address for the selected wallet.",
        "operationId": "generate_btc_address",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewBtcAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bitcoin Address Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BtcAddress"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"422 Unprocessable Entity\",\n    \"reason\": \"Validation failed: ...\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bitcoin Addresses"
        ],
        "summary": "Delete Bitcoin addresses",
        "description": "Deletes all Bitcoin addresses matching a filter. Returns the number deleted.",
        "operationId": "delete_btc_addresses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Total amount of results to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset where to start returning results",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "ids",
            "in": "query",
            "description": "List of IDs",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "wallet_id",
            "in": "query",
            "description": "Wallet ID. Automatically populated with your ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "Address",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "address_type",
            "in": "query",
            "description": "Status",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/BtcAddressType"
                }
              ]
            }
          },
          {
            "name": "used",
            "in": "query",
            "description": "Whether the address has been used",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "order_direction",
            "in": "query",
            "description": "Direction of the ordering of results",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      }
    },
    "/v1/bitcoin/addresses/{id}": {
      "get": {
        "tags": [
          "Bitcoin Addresses"
        ],
        "summary": "Find a Bitcoin address",
        "description": "Returns the Bitcoin address by its ID.",
        "operationId": "get_btc_address",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BtcAddress"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bitcoin Addresses"
        ],
        "summary": "Delete a Bitcoin address",
        "description": "Deletes a Bitcoin address by ID. Returns an empty body.",
        "operationId": "delete_btc_address",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BtcAddress": {
        "type": "object",
        "description": "Bitcoin Address",
        "required": [
          "id",
          "wallet_id",
          "address",
          "used",
          "address_type",
          "created_at"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Current deposit address"
          },
          "address_type": {
            "$ref": "#/components/schemas/BtcAddressType",
            "description": "Address type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date of creation in database"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Internal ID"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Date of update in database"
          },
          "used": {
            "type": "boolean",
            "description": "Whether the address has already been used on-chain"
          },
          "wallet_id": {
            "type": "string",
            "format": "uuid",
            "description": "Wallet ID"
          }
        }
      },
      "BtcAddressType": {
        "type": "string",
        "description": "Bitcoin address script type.",
        "enum": [
          "p2pkh",
          "p2sh",
          "p2wpkh",
          "p2tr"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Application error response",
        "required": [
          "status",
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Error reason",
            "example": "error message"
          },
          "status": {
            "type": "string",
            "description": "Error status",
            "example": "401 Unauthorized"
          }
        }
      },
...
      "NewBtcAddressRequest": {
        "type": "object",
        "description": "New Bitcoin Address Request",
        "properties": {
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BtcAddressType",
                "description": "Address type"
              }
            ]
          },
          "wallet_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Wallet ID to receive into. Required by admin endpoints; derived from the path on wallet-scoped endpoints."
          }
        }
      },
...
      "OrderDirection": {
        "type": "string",
        "description": "Direction of result ordering for list endpoints.",
        "enum": [
          "Asc",
          "Desc"
        ]
      },
...
    },
...
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
...
}